css/css-grid/grid-model/grid-container-scrollbars-sizing-002.html

<!DOCTYPE html>
<title>CSS Grid Layout Test: Intrinsic logical width with scrollbars</title>
<link rel="author" title="Sunil Ratnu" href="mailto:sunil.ratnu@samsung.com">
<link rel="help" href="https://drafts.csswg.org/css-grid/#overflow"/>
<link rel="stylesheet" href="/css/support/grid.css">
<link rel="stylesheet" href="/css/support/width-keyword-classes.css">
<meta name="assert" content="This test ensures that the grid container considers the scrollbars when computing the intrinsic logical widths."/>

<style>
.grid, .inline-grid {
    overflow: scroll;
}
.gridItem {
    width: 100px;
    height: 100px;
    background: green;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.grid'); checkLayout('.inline-grid');">

<div class='grid fit-content' data-expected-width='115'>
    <span class='gridItem'></span>
</div>
<div class='grid fit-content' style='left: 150px; overflow-y: hidden' data-expected-width='100'>
    <span class='gridItem'></span>
</div>
<div class='inline-grid fit-content' style='left: 300px' data-expected-width='115'>
    <span class='gridItem'></span>
</div>
<div class='inline-grid fit-content' style='left: 450px; overflow-y: hidden' data-expected-width='100'>
    <span class='gridItem'></span>
</div>
</body>
